+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # v137.0 (In progress)
 
+## 🦊 What's Changed 🦊
+
+### Glean
+- Updated to v63.1.0 ([#6584](https://github.com/mozilla/application-services/pull/6584))
+
 [Full Changelog](In progress)
 
 # v136.0 (_2025-02-03_)