+++ CHANGELOG.md
@@ -8,6 +8,16 @@
 ### Glean
 - Updated to v63.1.0 ([#6584](https://github.com/mozilla/application-services/pull/6584))
 
+### Logins
+New trait `NSSKeyManager`, which provides an NSS-backed key manager. Given a
+`PrimaryPasswordAuthenticator` implementation, the NSS keystore is used to
+store and retrieve the login encryption key. These features are only available
+when the Logins component is compiled with the `keydb` feature.  The
+`sync-pass` example has been adapted to use the NSSKeyManager. The example
+program can be called with an FX profile path in which the key is stored in the
+file key4.db and secured with a possibly set primary password.
+([#6571](https://github.com/mozilla/application-services/pull/6571))
+
 [Full Changelog](In progress)
 
 # v136.0 (_2025-02-03_)